Split models reproductions
